장비 보급시 소수점 입력이 가능하도록 설정하고,  siderbar에서 일부 기능을 수정하였다.

+++ app/views/equip/supplies-show.blade.php
@@ -73,14 +73,14 @@ $(function() {
 			@foreach ($types as $type)
 				var typeValue = $("#{{'count_'.$node->id.'_'.$type->id}}").text();
 				if (jQuery.isNumeric(typeValue)) {
-					sumNode += parseInt(typeValue);
+					sumNode += parseFloat(typeValue);
 				}
 				//노드별합계 계산
 			@endforeach
 			$("#{{'sum_node_'.$node->id}}").html('<b>'+sumNode+'</b>');
 			//노드별 합계를 넣었다.
 			//총 합계 계산
-			sumAll += parseInt(sumNode);
+			sumAll += parseFloat(sumNode);
 		@endforeach
 		$("#sum_all").html('<b>'+sumAll+'</b>');
 		//총 합계를 넣었다.
@@ -89,7 +89,7 @@ $(function() {
 			@foreach($lowerNodes as $node)
 				var nodeValue = $("#{{'count_'.$node->id.'_'.$type->id}}").text();
 				if(jQuery.isNumeric(nodeValue)){
-					sumType += parseInt(nodeValue);
+					sumType += parseFloat(nodeValue);
 				}
 			@endforeach
 			$("#{{'sum_type_'.$type->id}}").html('<b>'+sumType+'</b>');
